@AnthonyGrondin 2024-09-07
I've ran the `esp-mbedtls` self_tests with this patch, and there's a noticeable performance improvement after this upgrade. 
The variance of the results is also reduced across the board.

| Hash Algorithm | Software (cycles) | Hardware (cycles) | After upgrade |
|----------------|-------------------|-------------------|-------------------|
| SHA-1          |      3,390,785    |         896,889   |         744,096   |
| SHA-224        |      8,251,799    |         898,344   |         760,146   |
| SHA-256        |      8,237,932    |         901,709   |         766,755   |
| SHA-384        |     13,605,806    |         799,532   |         725,712   |
| SHA-512        |     13,588,104    |         801,556   |         725,315   |

I also ran a before / after test with `edge_server` and I haven't noticed a significant improvement / decline in metrics. It seems for a TLS operation, SHA time is an insignificant amount of the total processing time.

Before this upgrade:
```
App/part. size:    1,019,760/1,048,576 bytes, 97.25%

Benchmark 1: curl -k https://192.168.69.166
  Time (mean ± σ):      2.509 s ±  0.023 s    [User: 0.004 s, System: 0.003 s]
  Range (min … max):    2.469 s …  2.550 s    10 runs
```

After this upgrade:
```
App/part. size:    1,020,384/1,048,576 bytes, 97.31%

Benchmark 1: curl -k https://192.168.69.166
  Time (mean ± σ):      2.512 s ±  0.020 s    [User: 0.003 s, System: 0.004 s]
  Range (min … max):    2.489 s …  2.559 s    10 runs
```